// client2/src/lib/types.ts
export interface College {
  id: number;
  name: string;
  location: string | null;
  state: string | null;
  type: string | null;
  tuition: number | null;
  roomAndBoard: number | null;
  acceptanceRate: number | null;
  rating: number | null;
  size: string | null;
  rank: number | null;
  feesByIncome: string | null | {
    [key: string]: number;
  };
  usNewsTop150: number | null;
  bestLiberalArtsColleges: number | null;
  degreesAwardedPredominant: number | null;
  degreesAwardedHighest: number | null;
  admissionRateOverall: number | null;
  satScoresAverageOverall: number | null;
  pellGrantRate: number | null;
  completionRate4yr150nt: number | null;
  medianDebtCompletersOverall: number | null;
  medianDebtNoncompleters: number | null;
  demographicsMedianFamilyIncome: number | null;
  medianEarnings10yrsAfterEntry: number | null;
}